Removed setScaledOnDPI property; replaced with ability to completely customize injection HTML header & body code using the new setInjectionBodyCode(string) and setInjectionHeaderCode(String) properties (see commented out sample usage in ORMMABanner.java from samples/simple.)
Bump version string to 2.11beta3

Change on size changed to not stop timer if view state is default.

Hide virtual keyboard when updating or expanding ad view so that size is maintained correctly.

Make sure close button or region is always clickable in upper right corner region of ad view.


git-svn-id: http://mocean-sdk-android.googlecode.com/svn/trunk@156 f3a8e223-0ac0-4266-b396-771bc5b56a4e